requirements.txt
absl-py==0.15.0
aiohttp==3.8.6
aiosignal==1.3.1
appdirs==1.4.3
astunparse==1.6.2
async-timeout==4.0.3
atomicwrites==1.1.5
attrs==19.3.0
Automat==0.8.0
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.1
blinker==1.4
blosc==1.7.0
boto3==1.29.3
botocore==1.32.3
bottle==0.12.15
cachetools==4.0.0
caffe==1.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
charset-normalizer==3.3.0
Click==7.0
cloud-init==23.3.1
cloudpickle==1.3.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contextvars==2.4
cryptography==2.8
ctop==1.0.0
cycler==0.10.0
Cython==0.29.14
dask==2.8.1+dfsg
datasets==2.14.5
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.7
distlib==0.3.0
distro==1.4.0
distro-info==0.23+ubuntu1.1
docker==4.1.0
entrypoints==0.3
et-xmlfile==1.0.1
fail2ban==0.11.1
fasttext==0.9.1
filelock==3.0.12
flake8==3.7.9
flatbuffers==1.12
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.2
gast==0.4.0
Glances==3.1.3
google-auth==1.5.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.1
h5py==2.10.0
html5lib==1.0.1
httplib2==0.14.0
huggingface-hub==0.17.3
hyperlink==19.0.0
icdiff==1.9.5
idna==2.8
imageio==2.4.1
immutables==0.20
importlib-metadata==1.5.0
incremental==16.10.1
influxdb==5.2.0
iotop==0.6
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
javalang==0.13.0
jdcal==1.0
jedi==0.15.2
Jinja2==2.10.1
jmespath==1.0.1
joblib==1.3.2
jsonlines==4.0.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-console==6.0.0
jupyter-core==4.6.3
kaptan==0.5.10
keras==2.13.1
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libtmux==0.8.2
locket==0.2.0
lxml==4.5.0
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==4.2.0
mpi4py==3.0.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.4
networkx==2.4
nltk==3.8.1
nose==1.3.7
notebook==6.0.3
numexpr==2.7.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==7.352.0
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.3
opt-einsum==3.3.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.4.2
parameterized==0.7.0
parso==0.5.2
partd==1.0.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.0
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==4.21.12
psutil==5.5.1
py==1.8.1
py-cpuinfo==9.0.0
pyarrow==13.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pybind11==2.11.1
pycodestyle==2.5.0
pycparser==2.19
pycryptodomex==3.6.1
pycuda==2019.1.2
pydot==1.4.1
pyflakes==2.1.1
Pygments==2.3.1
PyGObject==3.36.0
pygpu==0.7.6
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyrsistent==0.15.5
pyserial==3.4
pysmi==0.3.2
pysnmp==4.4.6
pystache==0.5.4
pytest==4.6.9
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
pytools==2019.1.1
pytz==2023.3.post1
PyWavelets==0.5.1
PyYAML==5.3.1
pyzmq==18.1.1
regex==2023.10.3
requests==2.22.0
requests-oauthlib==1.0.0
requests-unixsocket==0.2.0
rsa==4.0
s3transfer==0.7.0
safetensors==0.4.0
scikit-cuda==0.5.3
scikit-image==0.16.2
scikit-learn==1.3.2
scikit-learn-extra==0.3.0
scipy==1.10.1
SecretStorage==2.3.1
Send2Trash==1.5.0
sentencepiece==0.1.99
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.5.6
soupsieve==1.9.5
ssh-import-id==5.10
sympy==1.12
systemd-python==234
tables==3.6.1
tensorboard==2.13.0
tensorflow==2.13.0
tensorflow-estimator==2.13.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.4
Theano==1.0.4
threadpoolctl==3.2.0
tmuxp==1.5.4
tokenizers==0.14.1
toolz==0.9.0
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==5.1.1
tqdm==4.66.1
traitlets==4.3.3
transformers==4.34.1
triton==2.1.0
Twisted==18.9.0
typing-extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36
unattended-upgrades==0.1
unidiff==0.7.5
urllib3==1.25.8
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.53.0
Werkzeug==0.16.1
widgetsnbextension==2.0.0
wrapt==1.11.2
xarray==2023.1.0
xlrd==1.1.0
xlwt==1.3.0
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.2
zipp==1.0.0
zope.interface==4.7.1
